MT#56447 support log retention
... in websocket test script even if the tests succeed, for debug
purposes.

+++ t/auto-daemon-tests-websocket.py
@@ -3317,7 +3317,7 @@ if __name__ == "__main__":
 
     eventloop.close()
 
-    if code == 0:
+    if code == 0 and not os.environ.get("RETAIN_LOGS"):
         os.unlink(so.name)
         os.unlink(se.name)
     else:
